requirements.txt
flask==2.0.1
flask-cors==3.0.10
PyJWT==2.3.0
web3==6.0.0
eth-account 